+++ Informatik/Informatik/projectile/Projectile.cpp
@@ -37,9 +37,9 @@ void Projectile::render(int xoff, int yoff)
     SDL_Rect src = {current_anim * 32, 0, 32, 32};
 
     if(max_anim != 1) renderWithShading(texture, src, r); // Animations cant be rotated...
-    else renderWithRotation(texture, src, r, -rotationAngle + PI / 4.0f, true);
+    else renderWithRotation(texture, src, r, -rotationAngle + (float) PI / 4.0f, true);
     
-    level->window->lights.addLight((float) r.x + data.width / 2.0f, (float) r.y + data.height / 2.0f, 5.0f, 0xFFFFA080, 0.9, 0.2);
+    level->window->lights.addLight((float) (r.x + data.width / 2.0f), (float) (r.y + data.height / 2.0f), 5.0f, 0xFFFFA080, 0.9f, 0.2f);
 }
 
 void Projectile::update(const uint8_t *keys)
